+++ pkg/types/boxes/externals.go
@@ -22,6 +22,9 @@ type BoxesFileMixings struct {
     // Legend definition used in this diagram
     Legend *Legend  `yaml:"legend,omitempty"`
 
+    // dictionary for layout mixins. key of the dictionary is the caption of the box that will take the additional content
+    LayoutMixins map[string]LayoutMixin  `yaml:"layoutMixins,omitempty"`
+
     // dictionary of connection objects
     Connections map[string]ConnectionCont  `yaml:"connections,omitempty"`
 
@@ -37,6 +40,7 @@ type BoxesFileMixings struct {
 func NewBoxesFileMixings() *BoxesFileMixings {
     return &BoxesFileMixings{
         Legend: NewLegend(),
+        LayoutMixins: make(map[string]LayoutMixin, 0),
         Connections: make(map[string]ConnectionCont, 0),
         Formats: make(map[string]Format, 0),
         FormatVariations: NewFormatVariations(),
@@ -52,6 +56,10 @@ func CopyBoxesFileMixings(src *BoxesFileMixings) *BoxesFileMixings {
     ret.Title = src.Title
     ret.Version = src.Version
     ret.Legend = CopyLegend(src.Legend)
+    ret.LayoutMixins = make(map[string]LayoutMixin, 0)
+    for k, v := range src.LayoutMixins {
+        ret.LayoutMixins[k] = v
+    }
     ret.Connections = make(map[string]ConnectionCont, 0)
     for k, v := range src.Connections {
         ret.Connections[k] = v
@@ -68,6 +76,9 @@ func CopyBoxesFileMixings(src *BoxesFileMixings) *BoxesFileMixings {
 
     return &ret
 }
+
+
+